Generalized special saving behavior for "exposed" properties
Instead, if the Mavo has no separate edit view, any changes to
properties are instantly (-ish. debounced to 1sec) saved. This was the
spirit of the previous code, but it was awfully specific and convoluted.
This also opens the way for dropping the whole “exposed property”
concept, which is confusing and convoluted.