Naïve fix for #93
If value is the same as the default when the default changes, then the
value updates to the default. This could be useful in some cases, but
really irritating when elements are involved that go through a range of
values (e.g. spinners and sliders), so it might happen that at some
point the value matches, so it gets tied to a default it never meant to
be.
The proper solution is to keep track whether the default is actually
used, and only use it when the user deletes the current value.