[major change] Add reference to node via symbol, flag to turn primitives into objects
Lays the groundwork for #81
Also, objects as primitives were already used for collections of
primitives and groups with a single property with the same name as the
group. This harmonizes these cases and makes sure native functions and
operators return consistent results whether they are passed primitives
or primitivish objects (yes, I totally just made up that word).