Important change: Descendant groups now have priority over ancestor groups in expression evaluation
The previous behavior resulted in some *really* unexpected results when
a property with the same name was present in multiple levels of
multiple nested collections. Consider this: collection.name is expected
to return all names from inside the collection. However, if collection
is inside a group with a name property, it was returning that one!!