Workaround for issue reported by @martypdx
Due to issue #30 .mv-empty added by Mavo interferes with expressions on
class and gets wiped. Due to the CSS depending on it to provide a
clickable area on edit, these properties were not editable with the
mouse.
Also, after some testing it looks like .mv-empty is not even needed
here, because these properties have mv-attribute=none and are not being
edited, so :empty should suffice.