Implement right and left unary and identity for binaryOperation() (#361)
As of now, the implementation is just a block of nested if statements checking if additional scalars or identities are implemented by the developer, but I'm looking for a less verbose way to do all this. Just hoping that my logic is going in the right direction though.

After that, I'm going to start changing all `identity` to `default` and also fix some weird behaviors with existing binary operations.