add button & bottom-up refactor, fixes #457
Also makes the code easier to understand, since lazy properties make it
impossible to tell where code is actually executed.
Introduces a new small regression: The add button in the item bar now
always inserts before the current item, regardless of mv-order. Once we
switch to actions for these, that will be fixed.