Fix one of the regressions in #687
Previously, all observation was paused on all primitives when one primitive needed to pause its observer. This caused actionable mutations to be dropped on other primitives, just because one primitive was doing something that needed to disable observation. We now switched to a flag on the primitive itself, which should fix any such issues.