Fix second regression in mv-value tests, closes #687
Because after the commit that introduced this, primitives were edited immediately, it meant that primitive.editor was updated, but no change was being made in the DOM, so the new editor just floated in DOM hyperspace while the old editor was still being displayed.
This could have theoretically caused a bug even before, but in much fewer cases.