Move data-mv-* to mv-* conversion to be global on the whole document
No need to run extra queries with every Mavo app construction.

This does have one downside beyond scoping: if anyone is using Mavo to manually initialize apps in iframes, any fixup that's not in the constructor will not work. However, a) that's very rare b) in those cases they usually include Mavo in the iframe too and c) this kind of fixup is only needed if they use data-mv-* attributes anyway, which is also quite rare.